Implement RoutingDatabase abstraction and refactor ruting code to make use of it.

+++ libosmscout/src/osmscout/ost/Scanner.cpp
@@ -129,7 +129,7 @@ void Buffer::SetPos(int value) {
 
 int Buffer::Read()
 {
-  if (bufPos < bufLen) {
+  if (bufPos < (int)bufLen) {
     return buf[bufPos++];
   } else {
     return EoF;
